% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contr.wec.weighted.R
\name{contr.wec.weighted}
\alias{contr.wec.weighted}
\title{Calculates contrasts for a weighted factor variable based on weighted effect coding}
\usage{
contr.wec.weighted(x, omitted, weights)
}
\arguments{
\item{x}{Grouping variable of class factor}

\item{omitted}{Label of the factor label that should be taken as the omitted category}

\item{weights}{Numeric vector of non-negative weights
from \code{x} as well as from \code{weights} prior to variance estimation.}
}
\value{
Returns a contrast matrix based on weighted effect coding.
}
\description{
Function works equivalent to \code{\link[wec]{contr.wec}} from the \code{wec} package, but allows
for weighted contrasts.
}
\examples{
### exemplary data according to wec paper
dat <- data.frame ( group = as.factor(c(rep(1,3), rep(2,2))), wgt = c(2/3, 4/3, 2, 3/8, 5/8))
### default contrasts
contrasts(dat[,"group"])
### weighted effect coding for weighted data
contr.wec.weighted(x= dat[,"group"], omitted=1,weights=dat[,"wgt"])
### equal to weighted effect coding: wec::contr.wec(x= dat[,"group"], omitted=1)
contr.wec.weighted(x= dat[,"group"], omitted=1,weights=rep(1, nrow(dat)))

}
