% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtdTable.r
\name{wtdTable}
\alias{wtdTable}
\title{Weighted frequency tables.}
\usage{
wtdTable(x, weights, na.rm = FALSE)
}
\arguments{
\item{x}{a character or category or factor vector}

\item{weights}{a numeric vector of non-negative weights}

\item{na.rm}{set to \code{FALSE} to suppress checking for NAs. If \code{TRUE}, NAs are removed
from \code{x} as well as from \code{weights} prior to variance estimation.}
}
\value{
a frequency table
}
\description{
Compute a weighted frequency table.
}
\details{
This function provides similar functionality as the \code{\link[Hmisc]{wtd.table}} function from \code{Hmisc}.
}
\examples{
x <- c(50, 1, 25)
w <- c(1, 4, 1)

# compute weighted variance
wtdTable(x, w)

}
