% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtdVar.r
\name{wtdVar}
\alias{wtdVar}
\title{Weighted variance.}
\usage{
wtdVar(x, weights, na.rm = FALSE)
}
\arguments{
\item{x}{numeric vector}

\item{weights}{a numeric vector of non-negative weights}

\item{na.rm}{set to \code{FALSE} to suppress checking for NAs. If \code{TRUE}, NAs are removed
from \code{x} as well as from \code{weights} prior to variance estimation.}
}
\value{
a scalar
}
\description{
Compute a weighted variance.
}
\details{
This function provides similar functionality as the \code{\link[Hmisc]{wtd.var}} function from \code{Hmisc}.
}
\examples{
x <- c(50, 1, 25)
w <- c(1, 4, 1)

# compute weighted variance
wtdVar(x, w)

}
